:wrench: :apple: Added borders to all buttons and increased size to 50px
- Moved `BuildFromHTML` to the only place it is used.
- Hide the title when showing First Debrief
